Add collisions between player and enemy characters
Collisions between the player and the enemy characters are possible. The
collision damages both the player and the enemy - substracts a small
portion of HP. After the collision, the player becomes immune for a
brief period so he does not lose his entire health pool in a 1/4 of a
second.